Add SQLite integration

+++ UI/ViewModels/MainViewModel.cs
@@ -14,6 +14,7 @@ using CORE;
 using Strzelecki_Baranowski.DuckApp.BL;
 using Strzelecki_Baranowski.DuckApp.INTERFACES;
 using Strzelecki_Baranowski.DuckApp.UI;
+using static System.Runtime.InteropServices.JavaScript.JSType;
 
 namespace Strzelecki_Baranowski.DuckApp.ViewModels
 {
@@ -53,7 +54,7 @@ namespace Strzelecki_Baranowski.DuckApp.ViewModels
         private IEnumerable _list;
 
         [ObservableProperty]
-        private object _selectedItem;
+        private object? _selectedItem;
 
         [ObservableProperty]
         private Visibility _editVisibility = Visibility.Collapsed;
@@ -138,7 +139,7 @@ namespace Strzelecki_Baranowski.DuckApp.ViewModels
             if (ListType == "ducks")
             {
                
-                var newduck  = new DuckViewModel(_businessLogic.GetNewDuck(), ProducerVM.Producers.FirstOrDefault(x=>x.ID==0)) ;
+                var newduck  = new DuckViewModel(_businessLogic.GetNewDuck(), /*ProducerVM.Producers.FirstOrDefault(x=>x.ID==0)*/null) ;
                 newduck.ID = -1;
                 newduck.Name = "";
                 DummyObject = newduck;
@@ -164,38 +165,72 @@ namespace Strzelecki_Baranowski.DuckApp.ViewModels
                     //var duck = SelectedItem as DuckViewModel;
                     //if (duck == null) return;
 
+                    var result = MessageBox.Show(
+                        $"Do you really want do delete duck '{duckVM.Name}'?",
+                        "Delete confirmation",
+                        MessageBoxButton.YesNo,
+                        MessageBoxImage.Warning); //TODO Tak i Nie  nie są po angielsku - co robić? własne okno?
+
+
+                    if (result == MessageBoxResult.No) return;
+
                     DuckVM.Ducks.Remove(/*SelectedItem as DuckViewModel*/duckVM);
                     _businessLogic.DeleteDuck(duckVM.ID);
 
                 }
                 else if (SelectedItem is ProducerViewModel prodVM)
                 {
-                    //var prodVM = SelectedItem as ProducerViewModel; //TODO przy usuwaniu przeba zmienić odniesienia we wszystkich kaczkach
-                    //if (prodVM == null) return;
 
-                    int defaultProducerId = 0;
+                    var result = MessageBox.Show(
+                        $"Do you really want do delete producer '{prodVM.Name}'?\n\nWarning: All ducks assigned to him will also be deleted!",
+                        "Delete confirmation",
+                        MessageBoxButton.YesNo,
+                        MessageBoxImage.Warning); //TODO Tak i Nie  nie są po angielsku - co robić? własne okno?
 
-                    var fallbackProducer = ProducerVM.Producers.FirstOrDefault(p => p.ID == defaultProducerId);
+                    if (result == MessageBoxResult.No) return;
 
-                 
-                    var ducksToUpdate = DuckVM.Ducks.Where(d => d.ProducerID == prodVM.ID).ToList();
+                    var ducksToRemove = DuckVM.Ducks.Where(d => d.ProducerID == prodVM.ID).ToList();
 
-                    foreach (var duck in ducksToUpdate)
+                    foreach (var duck in ducksToRemove)
                     {
-                        duck.ProducerVM = fallbackProducer;
-                        duck.ProducerID = fallbackProducer?.ID ?? defaultProducerId;
-                        duck.ProducerName = fallbackProducer?.Name ?? "Inny producent";
+                        DuckVM.Ducks.Remove(duck);
                     }
 
-                    ProducerVM.Producers.Remove(prodVM);
-                    _businessLogic.DeleteProducer(prodVM.ID);
-                }
 
+                    ProducerVM.Producers.Remove(prodVM);
+
+
+                    _businessLogic.DeleteProducer(prodVM.ID);
+
+                    SelectedItem = null;
+
+
+                    ////var prodVM = SelectedItem as ProducerViewModel; //TODO przy usuwaniu przeba zmienić odniesienia we wszystkich kaczkach
+                    ////if (prodVM == null) return;
+
+                    //int defaultProducerId = 0;
+
+                    //var fallbackProducer = ProducerVM.Producers.FirstOrDefault(p => p.ID == defaultProducerId);
+
+
+                    //var ducksToUpdate = DuckVM.Ducks.Where(d => d.ProducerID == prodVM.ID).ToList();
+
+                    //foreach (var duck in ducksToUpdate)
+                    //{
+                    //    duck.ProducerVM = fallbackProducer;
+                    //    duck.ProducerID = fallbackProducer?.ID ?? defaultProducerId;
+                    //    duck.ProducerName = fallbackProducer?.Name ?? "Inny producent";
+                    //}
+
+                    //ProducerVM.Producers.Remove(prodVM);
+                    //_businessLogic.DeleteProducer(prodVM.ID);
+                }
+                SelectedItem = null;
 
             }
-            catch
+            catch (Exception ex)
             {
-                //TODO ktoś coś
+                MessageBox.Show($"Error: {ex.Message}");
             }
         }
 
@@ -251,51 +286,79 @@ namespace Strzelecki_Baranowski.DuckApp.ViewModels
         {
             try
             {
+                
+                if (DummyObject is DuckViewModel duck)
+                {
+                    duck.Validate(); 
+                }
+                else if (DummyObject is ProducerViewModel producer)
+                {
+                    producer.Validate();
+                }
+
                 if (Bledy.Count() > 0)
+                {
+
+                    string komunikat = string.Join("\n ", Bledy);
+                    System.Windows.MessageBox.Show(
+                    $"Cannot save. The form contains errors:\n\n{komunikat}",
+                    "Validation error",
+                    System.Windows.MessageBoxButton.OK,
+                    System.Windows.MessageBoxImage.Warning);
                     return;
+
+                }
+
+
+
+
                 switch (DummyObject)
                 {
-                    
-                       
-
-                //if (SelectedItem == null) return;
-                //if (SelectedItem.GetType() == typeof(DuckViewModel))
-                //{
-
-                case DuckViewModel duckVM:
-                    var editedDuck = DummyObject as DuckViewModel;
-                    if (editedDuck == null) return;
-
-                    if (editedDuck.ID == -1)
-                    {
-                        int iD = _businessLogic.AddNewDuck(editedDuck.GetDuck()/*editedDuck.Name, editedDuck.ProducerID, editedDuck.Price, editedDuck.Photo, editedDuck.Description, editedDuck.Category */);
-                        
-                        editedDuck.ID = iD; //TODO trzebaby aktualizować IDuck _duck chyba że się go zupełnie pozbywamy
-                        editedDuck.ProducerID = editedDuck.ProducerVM!.ID;
-                        editedDuck.ProducerName = editedDuck.ProducerVM.Name;
-                        DuckVM.Ducks.Add( editedDuck );
-                        SelectedItem = editedDuck;
 
 
-                        } else
-                    {
 
+                    //if (SelectedItem == null) return;
+                    //if (SelectedItem.GetType() == typeof(DuckViewModel))
+                    //{
 
-                    
+                    case DuckViewModel editedDuck:
+                        //var editedDuck = DummyObject as DuckViewModel;
+                        //if (editedDuck == null) return;
 
-
-                        var ducks = DuckVM.Ducks;
-                        for (int i = 0; i < ducks.Count; i++)
+                        if (editedDuck.ID == -1)
                         {
-                            if (ducks[i].ID == editedDuck.ID)
+
+                            //editedDuck.ProducerID = editedDuck.ProducerVM!.ID;
+                            //editedDuck.ProducerName = editedDuck.ProducerVM.Name;
+
+                            int iD = _businessLogic.AddNewDuck(editedDuck.GetDuck()/*editedDuck.Name, editedDuck.ProducerID, editedDuck.Price, editedDuck.Photo, editedDuck.Description, editedDuck.Category */);
+
+                            editedDuck.ID = iD; 
+
+                            DuckVM.Ducks.Add(editedDuck);
+                            SelectedItem = editedDuck;
+
+
+                        }
+                        else
+                        {
+
+
+
+
+
+                            var ducks = DuckVM.Ducks;
+                            for (int i = 0; i < ducks.Count; i++)
                             {
+                                if (ducks[i].ID == editedDuck.ID)
+                                {
 
 
                                     ducks[i].UpdateFrom(editedDuck);
                                     SelectedItem = ducks[i];
                                     break;
+                                }
                             }
-                        }
                             _businessLogic.UpdateDuck(editedDuck.GetDuck()); //TODO nie przenosić przed UPDATEFrom bo wtdy viewmodel nie łapie zmian XD
 
 
@@ -308,43 +371,43 @@ namespace Strzelecki_Baranowski.DuckApp.ViewModels
 
                         break;
 
-                //}
-                //else
-                //{
-                case ProducerViewModel producerVM:
+                    //}
+                    //else
+                    //{
+                    case ProducerViewModel editedProd:
 
-                    var editedProd = DummyObject as ProducerViewModel;
-                    if (editedProd == null) return;
+                        //var editedProd = DummyObject as ProducerViewModel;
+                        //if (editedProd == null) return;
 
-                    if (editedProd.ID == -1)
-                    {
-                        int iD = _businessLogic.AddNewProducer(editedProd.Name, editedProd.Website);
+                        if (editedProd.ID == -1)
+                        {
+                            int iD = _businessLogic.AddNewProducer(editedProd.GetProd()/*editedProd.Name, editedProd.Website*/);
 
-                        editedProd.ID = iD; //TODO trzeba by aktualizować  _producer chyba że się go zupełnie pozbywamy (wtedy trzeba przerobić update w BLC)
+                            editedProd.ID = iD; //TODO trzeba by aktualizować  _producer chyba że się go zupełnie pozbywamy (wtedy trzeba przerobić update w BLC)
 
                             ProducerVM.Producers.Add(editedProd);
                             SelectedItem = editedProd;
 
                         }
                         else
-                    {
-
-                        
-
-
-
-                        var prods = ProducerVM.Producers;
-                        for (int i = 0; i < prods.Count; i++)
                         {
-                            if (prods[i].ID == editedProd.ID)
-                            {
-                                prods[i].UpdateFrom(editedProd);
-                                SelectedItem = prods[i];
-                                break;
-                            }
-                        }
 
-                        _businessLogic.UpdateProducer(editedProd.GetProd());
+
+
+
+
+                            var prods = ProducerVM.Producers;
+                            for (int i = 0; i < prods.Count; i++)
+                            {
+                                if (prods[i].ID == editedProd.ID)
+                                {
+                                    prods[i].UpdateFrom(editedProd);
+                                    SelectedItem = prods[i];
+                                    break;
+                                }
+                            }
+
+                            _businessLogic.UpdateProducer(editedProd.GetProd());
 
                         }
 
@@ -356,7 +419,7 @@ namespace Strzelecki_Baranowski.DuckApp.ViewModels
                 }
 
 
-               
+
 
 
                 DetailsVisibility = Visibility.Visible;
@@ -366,8 +429,11 @@ namespace Strzelecki_Baranowski.DuckApp.ViewModels
             catch (Exception ex)
             {
                 //TODO jakiś komunikat może?
-                throw ex;
+               MessageBox.Show("Saving error occurred: " + ex.Message);
+
+                //throw ex;
             }
+
         }
 
         public IEnumerable<string> Bledy
